pyproject.toml
[tool.poetry]
name = "base_auth_backend"
version = "0.1.0"
description = "Backend de autenticación multitenant con FastAPI"
authors = ["Arthur Monsalve"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.0"
pydantic = "^2.5.0"
PyYAML = "^6.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.10.1"
python-dotenv = "^1.1.1"
uvicorn = "^0.35.0"
email-validator = "^2.2.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
pytest = "^7.4.0"
isort = "^5.12.0"
alembic = "^1.16.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

